requirements.txt
black
pycodestyle
pyinstaller
pyserial
PyQt6
pyqt6-tools
pyvjoy
